Use Blueprint on booking function
-Separate api/booking using blueprint
-Isolate token.py, database_conn.py
